More on testbed 01.
A mechanism is provided to tell the StubFactory that we expect to see
a request for a RPC call.
When the call is actually made, the StubFactory can report its
arguments and the NodeID of the destination(s).
The testbed 01 calls make_connectivity. Then verifies that an ETP is
sent in order to report the address change to any neighbour.