Fix bugs.
A fingerprint was not correctly updated if its elderships changed
whilst its ID did not.
A destination of level 0 is always a real node (not a connectivity
one) so its 'nn' (number of nodes inside) is always 1. Instead, the
current node might well be a connectivity one. So its 'nn' must be
'my_nodes_inside[0]' instead of simply 1.