Changes and more tests in testbed 01.
Since a fix has been applied to proof-of-concept, the sequence of
events changed. We apply those changes to testbed01.
The fix to proof-of-concept was:
When we create the QspnManager of a new identity (both with enter_net
and migration) we have to pass the previous QspnManager while that
one hasn't yet done 'make_connectivity'.