More on testbeds.
Implement `expect_got_destroy` to check when we should do this RPC
 call.
Add call to `destroy` when dismissing an identity, and expect to see
 a RPC call `got_destroy`.
In testbed 01 when dismissing alfa0 we see a RPC call to nobody.
In testbed 01 when dismissing alfa1 we see a RPC call to beta0.
In testbed 02 when dismissing beta0 we see a RPC call to nobody.