Revisiting Qspn.
When removing arcs by initiative of the module, we always emit
a signal to the program.
The signal contains a boolean to state if the issue is with the
link, so that the program decides whether to remove the identity-arc
or the physical-arc.
When removing arcs for errors print a warning, instead of relaying
a message through the signal.
Removed old testbeds.